:card_file_box: use bigint as id

+++ db/migrations/20230624020202_create-settings.sql
@@ -1,8 +1,8 @@
 -- migrate:up
 CREATE TABLE settings(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     "key" VARCHAR(255) NOT NULL,
-    user_id INT,
+    user_id BIGINT,
     salt BYTEA,
     "value" BYTEA NOT NULL,
     "version" INT NOT NULL DEFAULT 0,
