+++ db/migrations/20230624022122_create-tags.sql
@@ -1,6 +1,6 @@
 -- migrate:up
 CREATE TABLE tags(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     code VARCHAR(63) NOT NULL,
     sort_order INT NOT NULL DEFAULT 100,
     deleted_at TIMESTAMP WITHOUT TIME ZONE,
@@ -10,10 +10,10 @@ CREATE TABLE tags(
 );
 CREATE UNIQUE INDEX idx_tags_code ON tags(code);
 CREATE TABLE tag_resources(
-    id SERIAL PRIMARY KEY,
-    tag_id INT NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    tag_id BIGINT NOT NULL,
     resource_type VARCHAR(255) NOT NULL,
-    resource_id INT NOT NULL,
+    resource_id BIGINT NOT NULL,
     sort_order INT NOT NULL DEFAULT 100,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
