+++ db/migrations/20230624022128_create-categories.sql
@@ -1,6 +1,6 @@
 -- migrate:up
 CREATE TABLE categories(
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     code VARCHAR(63) NOT NULL,
     "left" INT NOT NULL,
     "right" INT NOT NULL,
@@ -12,10 +12,10 @@ CREATE TABLE categories(
 CREATE UNIQUE INDEX idx_categories_code ON categories(code);
 CREATE UNIQUE INDEX idx_categories_left_right ON categories("left", "right");
 CREATE TABLE category_resources(
-    id SERIAL PRIMARY KEY,
-    category_id INT NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    category_id BIGINT NOT NULL,
     resource_type VARCHAR(255) NOT NULL,
-    resource_id INT NOT NULL,
+    resource_id BIGINT NOT NULL,
     sort_order INT NOT NULL DEFAULT 100,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
